updated NHDNetwork.py with all changes for use with current troute methods
